Refactored main.dart, added screens and Firebase integration
