+++ lib/firebase_options.dart
@@ -1,0 +1,86 @@
+// File generated by FlutterFire CLI.
+// ignore_for_file: type=lint
+import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
+import 'package:flutter/foundation.dart'
+    show defaultTargetPlatform, kIsWeb, TargetPlatform;
+
+/// Default [FirebaseOptions] for use with your Firebase apps.
+///
+/// Example:
+/// ```dart
+/// import 'firebase_options.dart';
+/// // ...
+/// await Firebase.initializeApp(
+///   options: DefaultFirebaseOptions.currentPlatform,
+/// );
+/// ```
+class DefaultFirebaseOptions {
+  static FirebaseOptions get currentPlatform {
+    if (kIsWeb) {
+      return web;
+    }
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+        return android;
+      case TargetPlatform.iOS:
+        return ios;
+      case TargetPlatform.macOS:
+        return macos;
+      case TargetPlatform.windows:
+        return windows;
+      case TargetPlatform.linux:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for linux - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      default:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions are not supported for this platform.',
+        );
+    }
+  }
+
+  static const FirebaseOptions web = FirebaseOptions(
+    apiKey: 'AIzaSyCEix764hsUdcEfWOdUB82Q6-qt7QMqxD0',
+    appId: '1:689918294994:web:39a232dae481d821ac260b',
+    messagingSenderId: '689918294994',
+    projectId: 'lms-web-5096a',
+    authDomain: 'lms-web-5096a.firebaseapp.com',
+    storageBucket: 'lms-web-5096a.firebasestorage.app',
+  );
+
+  static const FirebaseOptions android = FirebaseOptions(
+    apiKey: 'AIzaSyAKFUREFJBXGO7FT9y7FTMTqQUowGpgL_Y',
+    appId: '1:689918294994:android:83bb9d66aff4734eac260b',
+    messagingSenderId: '689918294994',
+    projectId: 'lms-web-5096a',
+    storageBucket: 'lms-web-5096a.firebasestorage.app',
+  );
+
+  static const FirebaseOptions ios = FirebaseOptions(
+    apiKey: 'AIzaSyAAf3FLNLHHmNbzHob-PfB1bB8TcDJDjnY',
+    appId: '1:689918294994:ios:cdda0beff1d19410ac260b',
+    messagingSenderId: '689918294994',
+    projectId: 'lms-web-5096a',
+    storageBucket: 'lms-web-5096a.firebasestorage.app',
+    iosBundleId: 'com.example.lms',
+  );
+
+  static const FirebaseOptions macos = FirebaseOptions(
+    apiKey: 'AIzaSyAAf3FLNLHHmNbzHob-PfB1bB8TcDJDjnY',
+    appId: '1:689918294994:ios:cdda0beff1d19410ac260b',
+    messagingSenderId: '689918294994',
+    projectId: 'lms-web-5096a',
+    storageBucket: 'lms-web-5096a.firebasestorage.app',
+    iosBundleId: 'com.example.lms',
+  );
+
+  static const FirebaseOptions windows = FirebaseOptions(
+    apiKey: 'AIzaSyCEix764hsUdcEfWOdUB82Q6-qt7QMqxD0',
+    appId: '1:689918294994:web:154e7a98360af3c5ac260b',
+    messagingSenderId: '689918294994',
+    projectId: 'lms-web-5096a',
+    authDomain: 'lms-web-5096a.firebaseapp.com',
+    storageBucket: 'lms-web-5096a.firebasestorage.app',
+  );
+}
