+++ lib/main.dart
@@ -1,7 +1,22 @@
 import 'package:flutter/material.dart';
-
-void main() {
-  runApp(const CollegeNotesApp());
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'auth_screen.dart';
+import 'firebase_options.dart';
+import 'screens/home_screen.dart';
+import 'screens/notes_screen.dart';
+import 'screens/opportunities_screen.dart';
+import 'screens/profile_screen.dart';
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  try {
+    await Firebase.initializeApp(
+      options: DefaultFirebaseOptions.currentPlatform,
+    );
+    runApp(const CollegeNotesApp());
+  } catch (e) {
+    print("Firebase Init Error: $e");
+  }
 }
 
 class CollegeNotesApp extends StatelessWidget {
@@ -19,8 +34,21 @@ class CollegeNotesApp extends StatelessWidget {
         useMaterial3: true,
         fontFamily: 'Roboto',
       ),
-      home: const MainScreen(),
       debugShowCheckedModeBanner: false,
+      home: StreamBuilder<User?>(
+        stream: FirebaseAuth.instance.authStateChanges(),
+        builder: (context, snapshot) {
+          if (snapshot.connectionState == ConnectionState.waiting) {
+            return const Scaffold(
+              body: Center(child: CircularProgressIndicator()),
+            );
+          } else if (snapshot.hasData) {
+            return const MainScreen(); // already logged in
+          } else {
+            return const AuthScreen(); // not logged in
+          }
+        },
+      ),
     );
   }
 }
@@ -105,739 +133,3 @@ class _MainScreenState extends State {
   }
 }
 
-class HomeScreen extends StatelessWidget {
-  const HomeScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('College Hub'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.search),
-            onPressed: () {},
-          ),
-          IconButton(
-            icon: const Icon(Icons.notifications_outlined),
-            onPressed: () {},
-          ),
-        ],
-      ),
-      body: SingleChildScrollView(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // Welcome Card
-            Card(
-              child: Container(
-                width: double.infinity,
-                padding: const EdgeInsets.all(20),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      'Welcome Back!',
-                      style: Theme.of(context).textTheme.headlineSmall,
-                    ),
-                    const SizedBox(height: 8),
-                    Text(
-                      'Share knowledge, discover opportunities',
-                      style: Theme.of(context).textTheme.bodyMedium,
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            const SizedBox(height: 24),
-
-            // Quick Stats
-            Row(
-              children: [
-                Expanded(
-                  child: _buildStatCard(
-                    context,
-                    'Notes Shared',
-                    '1,234',
-                    Icons.note,
-                    Colors.blue,
-                  ),
-                ),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: _buildStatCard(
-                    context,
-                    'Opportunities',
-                    '89',
-                    Icons.work,
-                    Colors.green,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 24),
-
-            // Recent Notes Section
-            Text(
-              'Recent Notes',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            const SizedBox(height: 16),
-            ListView.builder(
-              shrinkWrap: true,
-              physics: const NeverScrollableScrollPhysics(),
-              itemCount: 3,
-              itemBuilder: (context, index) {
-                return _buildNoteCard(context, index);
-              },
-            ),
-            const SizedBox(height: 24),
-
-            // Recent Opportunities Section
-            Text(
-              'Latest Opportunities',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            const SizedBox(height: 16),
-            ListView.builder(
-              shrinkWrap: true,
-              physics: const NeverScrollableScrollPhysics(),
-              itemCount: 2,
-              itemBuilder: (context, index) {
-                return _buildOpportunityCard(context, index);
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildStatCard(BuildContext context, String title, String value,
-                       IconData icon, Color color) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          children: [
-            Icon(icon, size: 32, color: color),
-            const SizedBox(height: 8),
-            Text(
-              value,
-              style: Theme.of(context).textTheme.headlineMedium?.copyWith(
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            Text(
-              title,
-              style: Theme.of(context).textTheme.bodySmall,
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildNoteCard(BuildContext context, int index) {
-    final notes = [
-      {'title': 'Data Structures - Chapter 5', 'subject': 'Computer Science', 'author': 'g Doe'},
-      {'title': 'Calculus Integration Notes', 'subject': 'Mathematics', 'author': 'Jane Smith'},
-      {'title': 'Physics Mechanics Summary', 'subject': 'Physics', 'author': 'Mike Johnson'},
-    ];
-
-    return Card(
-      margin: const EdgeInsets.only(bottom: 12),
-      child: ListTile(
-        leading: CircleAvatar(
-          backgroundColor: Theme.of(context).colorScheme.primary,
-          child: const Icon(Icons.note, color: Colors.white),
-        ),
-        title: Text(notes[index]['title']!),
-        subtitle: Text('${notes[index]['subject']} • ${notes[index]['author']}'),
-        trailing: IconButton(
-          icon: const Icon(Icons.bookmark_border),
-          onPressed: () {},
-        ),
-      ),
-    );
-  }
-
-  Widget _buildOpportunityCard(BuildContext context, int index) {
-    final opportunities = [
-      {'title': 'Summer Internship at Tech Corp', 'type': 'Internship', 'deadline': '2 days left'},
-      {'title': 'Research Assistant Position', 'type': 'Job', 'deadline': '1 week left'},
-    ];
-
-    return Card(
-      margin: const EdgeInsets.only(bottom: 12),
-      child: ListTile(
-        leading: CircleAvatar(
-          backgroundColor: Theme.of(context).colorScheme.secondary,
-          child: const Icon(Icons.work, color: Colors.white),
-        ),
-        title: Text(opportunities[index]['title']!),
-        subtitle: Text('${opportunities[index]['type']} • ${opportunities[index]['deadline']}'),
-        trailing: IconButton(
-          icon: const Icon(Icons.arrow_forward_ios),
-          onPressed: () {},
-        ),
-      ),
-    );
-  }
-}
-
-class NotesScreen extends StatefulWidget {
-  const NotesScreen({super.key});
-
-  @override
-  State createState() => _NotesScreenState();
-}
-
-class _NotesScreenState extends State {
-  String selectedCategory = 'All';
-  final List categories = ['All', 'Computer Science', 'Mathematics', 'Physics', 'Chemistry', 'Biology'];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Notes'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.search),
-            onPressed: () {},
-          ),
-          IconButton(
-            icon: const Icon(Icons.filter_list),
-            onPressed: () {},
-          ),
-        ],
-      ),
-      body: Column(
-        children: [
-          // Category Filter
-          Container(
-            height: 60,
-            padding: const EdgeInsets.symmetric(vertical: 8),
-            child: ListView.builder(
-              scrollDirection: Axis.horizontal,
-              itemCount: categories.length,
-              itemBuilder: (context, index) {
-                return Padding(
-                  padding: EdgeInsets.only(
-                    left: index == 0 ? 16 : 8,
-                    right: index == categories.length - 1 ? 16 : 0,
-                  ),
-                  child: FilterChip(
-                    label: Text(categories[index]),
-                    selected: selectedCategory == categories[index],
-                    onSelected: (selected) {
-                      setState(() {
-                        selectedCategory = categories[index];
-                      });
-                    },
-                  ),
-                );
-              },
-            ),
-          ),
-
-          // Notes List
-          Expanded(
-            child: ListView.builder(
-              padding: const EdgeInsets.all(16),
-              itemCount: 10,
-              itemBuilder: (context, index) {
-                return _buildNoteCard(context, index);
-              },
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildNoteCard(BuildContext context, int index) {
-    final notes = [
-      {
-        'title': 'Advanced Data Structures',
-        'subject': 'Computer Science',
-        'author': 'Sarah Wilson',
-        'rating': 4.8,
-        'downloads': 234,
-        'date': '2 days ago'
-      },
-      {
-        'title': 'Calculus II - Integration Techniques',
-        'subject': 'Mathematics',
-        'author': 'Robert Chen',
-        'rating': 4.6,
-        'downloads': 156,
-        'date': '5 days ago'
-      },
-      {
-        'title': 'Quantum Mechanics Basics',
-        'subject': 'Physics',
-        'author': 'Emily Davis',
-        'rating': 4.9,
-        'downloads': 89,
-        'date': '1 week ago'
-      },
-    ];
-
-    final note = notes[index % notes.length];
-
-    return Card(
-      margin: const EdgeInsets.only(bottom: 16),
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Expanded(
-                  child: Text(
-                    note['title'] as String,
-                    style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ),
-                IconButton(
-                  icon: const Icon(Icons.more_vert),
-                  onPressed: () {},
-                ),
-              ],
-            ),
-            const SizedBox(height: 8),
-            Row(
-              children: [
-                Chip(
-                  label: Text(
-                    note['subject'] as String,
-                    style: const TextStyle(fontSize: 12),
-                  ),
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                const SizedBox(width: 8),
-                Text(
-                  'by ${note['author']}',
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Icon(Icons.star, size: 16, color: Colors.amber),
-                const SizedBox(width: 4),
-                Text('${note['rating']}'),
-                const SizedBox(width: 16),
-                Icon(Icons.download, size: 16, color: Colors.grey[600]),
-                const SizedBox(width: 4),
-                Text('${note['downloads']}'),
-                const Spacer(),
-                Text(
-                  note['date'] as String,
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Expanded(
-                  child: OutlinedButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.visibility),
-                    label: const Text('Preview'),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Expanded(
-                  child: FilledButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.download),
-                    label: const Text('Download'),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class OpportunitiesScreen extends StatefulWidget {
-  const OpportunitiesScreen({super.key});
-
-  @override
-  State createState() => _OpportunitiesScreenState();
-}
-
-class _OpportunitiesScreenState extends State {
-  String selectedType = 'All';
-  final List types = ['All', 'Internship', 'Job', 'Research', 'Event', 'Workshop'];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Opportunities'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.search),
-            onPressed: () {},
-          ),
-          IconButton(
-            icon: const Icon(Icons.filter_list),
-            onPressed: () {},
-          ),
-        ],
-      ),
-      body: Column(
-        children: [
-          // Type Filter
-          Container(
-            height: 60,
-            padding: const EdgeInsets.symmetric(vertical: 8),
-            child: ListView.builder(
-              scrollDirection: Axis.horizontal,
-              itemCount: types.length,
-              itemBuilder: (context, index) {
-                return Padding(
-                  padding: EdgeInsets.only(
-                    left: index == 0 ? 16 : 8,
-                    right: index == types.length - 1 ? 16 : 0,
-                  ),
-                  child: FilterChip(
-                    label: Text(types[index]),
-                    selected: selectedType == types[index],
-                    onSelected: (selected) {
-                      setState(() {
-                        selectedType = types[index];
-                      });
-                    },
-                  ),
-                );
-              },
-            ),
-          ),
-
-          // Opportunities List
-          Expanded(
-            child: ListView.builder(
-              padding: const EdgeInsets.all(16),
-              itemCount: 8,
-              itemBuilder: (context, index) {
-                return _buildOpportunityCard(context, index);
-              },
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildOpportunityCard(BuildContext context, int index) {
-    final opportunities = [
-      {
-        'title': 'Software Engineering Internship',
-        'company': 'Tech Innovators Inc.',
-        'type': 'Internship',
-        'location': 'San Francisco, CA',
-        'deadline': '3 days left',
-        'salary': '\$3,000/month',
-        'urgent': true,
-      },
-      {
-        'title': 'Data Science Research Assistant',
-        'company': 'University Research Lab',
-        'type': 'Research',
-        'location': 'Cambridge, MA',
-        'deadline': '2 weeks left',
-        'salary': '\$2,500/month',
-        'urgent': false,
-      },
-      {
-        'title': 'Machine Learning Workshop',
-        'company': 'AI Academy',
-        'type': 'Workshop',
-        'location': 'Online',
-        'deadline': '5 days left',
-        'salary': 'Free',
-        'urgent': false,
-      },
-    ];
-
-    final opportunity = opportunities[index % opportunities.length];
-
-    return Card(
-      margin: const EdgeInsets.only(bottom: 16),
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        opportunity['title'] as String,
-                        style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 4),
-                      Text(
-                        opportunity['company'] as String,
-                        style: Theme.of(context).textTheme.bodyMedium,
-                      ),
-                    ],
-                  ),
-                ),
-                if (opportunity['urgent'] as bool)
-                  Container(
-                    padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                    decoration: BoxDecoration(
-                      color: Colors.red,
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                    child: const Text(
-                      'URGENT',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 10,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Chip(
-                  label: Text(
-                    opportunity['type'] as String,
-                    style: const TextStyle(fontSize: 12),
-                  ),
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                const SizedBox(width: 8),
-                Icon(Icons.location_on, size: 16, color: Colors.grey[600]),
-                const SizedBox(width: 4),
-                Text(
-                  opportunity['location'] as String,
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Icon(Icons.attach_money, size: 16, color: Colors.green),
-                const SizedBox(width: 4),
-                Text(
-                  opportunity['salary'] as String,
-                  style: const TextStyle(
-                    fontWeight: FontWeight.bold,
-                    color: Colors.green,
-                  ),
-                ),
-                const Spacer(),
-                Icon(Icons.access_time, size: 16, color: Colors.orange),
-                const SizedBox(width: 4),
-                Text(
-                  opportunity['deadline'] as String,
-                  style: const TextStyle(
-                    fontWeight: FontWeight.bold,
-                    color: Colors.orange,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Expanded(
-                  child: OutlinedButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.info),
-                    label: const Text('Details'),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Expanded(
-                  child: FilledButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.send),
-                    label: const Text('Apply'),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class ProfileScreen extends StatelessWidget {
-  const ProfileScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Profile'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.edit),
-            onPressed: () {},
-          ),
-        ],
-      ),
-      body: SingleChildScrollView(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          children: [
-            // Profile Header
-            Card(
-              child: Padding(
-                padding: const EdgeInsets.all(20),
-                child: Column(
-                  children: [
-                    CircleAvatar(
-                      radius: 50,
-                      backgroundColor: Theme.of(context).colorScheme.primary,
-                      child: const Icon(Icons.person, size: 50, color: Colors.white),
-                    ),
-                    const SizedBox(height: 16),
-                    Text(
-                      'John Doe',
-                      style: Theme.of(context).textTheme.headlineSmall,
-                    ),
-                    const SizedBox(height: 8),
-                    Text(
-                      'Computer Science • Final Year',
-                      style: Theme.of(context).textTheme.bodyMedium,
-                    ),
-                    const SizedBox(height: 4),
-                    Text(
-                      'john.doe@college.edu',
-                      style: Theme.of(context).textTheme.bodySmall,
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            const SizedBox(height: 20),
-
-            // Stats Cards
-            Row(
-              children: [
-                Expanded(
-                  child: _buildStatCard(context, 'Notes Shared', '12', Icons.note),
-                ),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: _buildStatCard(context, 'Downloads', '456', Icons.download),
-                ),
-              ],
-            ),
-            const SizedBox(height: 16),
-            Row(
-              children: [
-                Expanded(
-                  child: _buildStatCard(context, 'Applications', '8', Icons.send),
-                ),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: _buildStatCard(context, 'Bookmarks', '34', Icons.bookmark),
-                ),
-              ],
-            ),
-            const SizedBox(height: 20),
-
-            // Menu Items
-            Card(
-              child: Column(
-                children: [
-                  _buildMenuItem(context, 'My Notes', Icons.note, () {}),
-                  _buildMenuItem(context, 'My Applications', Icons.work, () {}),
-                  _buildMenuItem(context, 'Bookmarks', Icons.bookmark, () {}),
-                  _buildMenuItem(context, 'Settings', Icons.settings, () {}),
-                  _buildMenuItem(context, 'Help & Support', Icons.help, () {}),
-                  _buildMenuItem(context, 'About', Icons.info, () {}),
-                ],
-              ),
-            ),
-            const SizedBox(height: 20),
-
-            // Logout Button
-            SizedBox(
-              width: double.infinity,
-              child: OutlinedButton.icon(
-                onPressed: () {},
-                icon: const Icon(Icons.logout),
-                label: const Text('Logout'),
-                style: OutlinedButton.styleFrom(
-                  foregroundColor: Colors.red,
-                  side: const BorderSide(color: Colors.red),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildStatCard(BuildContext context, String title, String value, IconData icon) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          children: [
-            Icon(icon, size: 32, color: Theme.of(context).colorScheme.primary),
-            const SizedBox(height: 8),
-            Text(
-              value,
-              style: Theme.of(context).textTheme.headlineMedium?.copyWith(
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            Text(
-              title,
-              style: Theme.of(context).textTheme.bodySmall,
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildMenuItem(BuildContext context, String title, IconData icon, VoidCallback onTap) {
-    return ListTile(
-      leading: Icon(icon),
-      title: Text(title),
-      trailing: const Icon(Icons.arrow_forward_ios),
-      onTap: onTap,
-    );
-  }
-}
